refactor (tests): move CronExpression tests into correct test class.
